completed order checkout and premium checkout
