+++ src/components/stripe/checkoutForm/index.js
@@ -1,0 +1,1 @@
+export * from './checkoutForm.jsx'
