+++ src/components/stripe/elementsWrapper/index.js
@@ -1,0 +1,1 @@
+export * from './stripe.jsx'
