--- src/components/stripe/index.js
@@ -1,1 +1,0 @@
-export * from './stripe'
