+++ src/pages/checkout/index.js
@@ -1,0 +1,1 @@
+export * from "./checkout.jsx";
