--- src/pages/checkoutForm/index.js
@@ -1,1 +1,0 @@
-export * from "./checkoutForm.jsx";
