+++ src/pages/premium/index.js
@@ -1,0 +1,1 @@
+export * from "./premium.jsx";
