+++ src/utils/Router.jsx
@@ -16,7 +16,8 @@ import {PrivateRoute} from "./PrivateRoute.jsx";
 import {ViewProduct} from "../pages/viewProduct/index.js";
 import {Wishlist} from "../pages/wishlist/index.js";
 import {Orders} from "../pages/orders/index.js";
-import {Stripe} from "../components/stripe/index.js";
+import {Checkout} from "../pages/checkout/index.js";
+import {PremiumCheckout} from "../pages/premium/index.js";
 
 
 export default function Router() {
@@ -42,7 +43,8 @@ export default function Router() {
                 {path: "/profile", element: <Navigate to={"/"}/>},
                 {path: "/profile/edit", element: <PrivateRoute component={EditProfile}/>},
                 {path:"/cart", element: <PrivateRoute component={Cart}/>},
-                {path:"/checkout", element: <PrivateRoute component={Stripe}/>},
+                {path:"/checkout", element: <PrivateRoute component={Checkout}/>},
+                {path:"/premium", element: <PrivateRoute component={PremiumCheckout}/>},
                 {path: "/wishlist", element: <PrivateRoute component={Wishlist}/>},
                 {path: "/orders", element: <PrivateRoute component={Orders}/>},
             ]
